Auto-determine kernel and initrd locations
The stiefel-server script, when generating boot.tar, after mounting the boot
partition, now reads the GRUB or Syslinux configuration to determine which
kernel and initrd blobs to pack into boot.tar.
Alternatively, a stiefelsystem.json can be created on the boot partition.

This eliminates the need for the kernel and initrd cmdline parameters
for the stiefelsystem server (as started through stiefel-autokexec).